+++ src/python/DAS/core/das_aggregators.py
@@ -24,34 +24,62 @@ class ResultObject(object):
     """
     def __init__(self):
         self.result = 'Not available'
-    def sum(self, value):
+        self._id = None
+        self.rec_count = 0
+    def sum(self, obj):
         """Sum function for this object"""
+        value, _ = obj
         if  self.result == 'Not available':
             self.result = 0
         self.result += value
-    def count(self, value):
+        self.rec_count += 1
+    def count(self, obj):
         """Count function for this object"""
+        value, _ = obj
         if  self.result == 'Not available':
             self.result = 0
         self.result += 1
-    def max(self, value):
+        self.rec_count += 1
+    def max(self, obj):
         """Max function for this object"""
+        value, _id = obj
         if  self.result == 'Not available':
             self.result = 0
         if  value > self.result:
             self.result = value
-    def min(self, value):
+            if  _id:
+                self._id = _id
+        self.rec_count += 1
+    def min(self, obj):
         """Min function for this object"""
+        value, _id = obj
         if  self.result == 'Not available':
             self.result = 99999999999
         if  value < self.result:
             self.result = value
-    def avg(self, value):
-        """Average function for this object"""
-        pass
-#    def median(self, value):
-#        """Median function for this object"""
-#        pass
+            if  _id:
+                self._id = _id
+        self.rec_count += 1
+    def avg(self, obj):
+        """
+        Average function for this object, we store both sum and rec_count, such
+        that avg = sum/rec_count
+        """
+        value, _ = obj
+        if  self.result == 'Not available':
+            self.result = 0
+        self.result += float(value)
+        self.rec_count += 1
+    def median(self, obj):
+        """
+        Median function for this object, we store all values as array, such
+        that median can be found as middle value of array.
+        """
+        value, _ = obj
+        if  self.result == 'Not available':
+            self.result = []
+        self.result.append(value)
+        self.rec_count += 1
     
 def coroutine(func):
     """Coroutine helper, to be used as decorator"""
@@ -66,26 +94,27 @@ def decomposer(idict, target):
     """
     Receive input dict and pass its key,value pairs further to the target.
     """
+    _id = idict.get('_id', None)
     try:
-        target.send(idict)
+        target.send((idict, _id))
     except StopIteration:
         pass
 
 @coroutine
 def selector(key, target):
     """A coroutine that selector the recived dict for given key"""
-    row = (yield)
+    row, _id = (yield)
     if  dict_type(row):
         if  row.has_key(key):
             try:
-                target.send(row[key])
+                target.send((row[key], _id))
             except StopIteration:
                 pass
     elif isinstance(row, list):
         for item in row:
             if  dict_type(item) and item.has_key(key):
                 try:
-                    target.send(item[key])
+                    target.send((item[key], _id))
                 except:
                     pass
 
@@ -140,6 +169,53 @@ def das_func(func, ckey, genrows):
         obj  = compile(code, '<string>', 'exec')
         eval(obj)
     if  robj.result != 'Not available':
-        return robj.result
+        return robj
     return None
+
+def count(key, rows):
+    """DAS count aggregator function"""
+    robj = das_func('count', key, rows)
+    return {'value': robj.result}
+
+def das_avg(key, rows):
+    """DAS avg aggregator function"""
+    robj = das_func('avg', key, rows)
+    if  robj.rec_count:
+        data = {'value': float(robj.result)/robj.rec_count}
+    else:
+        data = {'value': 'N/A'}
+    return data
+
+def das_median(key, rows):
+    """DAS median aggregator function"""
+    robj = das_func('median', key, rows)
+    if  len(robj.result) == 1:
+        data = {'value': robj.result[0]}
+    if  len(robj.result) % 2:
+        data = {'value': robj.result[len(robj.result)/2]}
+    else:
+        val  = (robj.result[len(robj.result)/2-1] + \
+                robj.result[len(robj.result)/2] )/2
+        data = {'value': val}
+    return data
+
+def das_min(key, rows):
+    """DAS min aggregator function"""
+    robj = das_func('min', key, rows)
+    return {'value': robj.result, '_id': robj._id}
+
+def das_max(key, rows):
+    """DAS max aggregator function"""
+    robj = das_func('max', key, rows)
+    return {'value': robj.result, '_id': robj._id}
+
+def das_sum(key, rows):
+    """DAS sum aggregator function"""
+    robj = das_func('sum', key, rows)
+    return {'value': robj.result}
+
+def das_count(key, rows):
+    """DAS count aggregator function"""
+    robj = das_func('count', key, rows)
+    return {'value': robj.result}