+++ test/das_core_t.py
@@ -55,7 +55,8 @@ class testDASCore(unittest.TestCase):
         result = self.das.call(query)
         result = self.das.get_from_cache(query)
         result = [r for r in result][0]
-        expect = {"function": "count", "result": 5, "key": "zip.Placemark.address", "_id":0}
+        expect = {"function": "count", "result": {"value": 5}, 
+                  "key": "zip.Placemark.address", "_id":0}
         self.assertEqual(expect, result)
 
     def testIPService(self):